+++ rasa_core/agent.py
@@ -3,26 +3,30 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-import os
-
 import logging
+import os
+import shutil
+
+import typing
 from six import string_types
 from typing import Text, List, Optional, Callable, Any, Dict, Union
 
+from rasa_core import training
 from rasa_core.channels import UserMessage, InputChannel, OutputChannel
-from rasa_core.domain import TemplateDomain, Domain
+from rasa_core.domain import TemplateDomain, Domain, check_domain_sanity
 from rasa_core.events import Event
-from rasa_core.featurizers import Featurizer, BinaryFeaturizer
 from rasa_core.interpreter import NaturalLanguageInterpreter
-from rasa_core.policies import PolicyTrainer, Policy
+from rasa_core.policies import Policy
 from rasa_core.policies.ensemble import SimplePolicyEnsemble, PolicyEnsemble
 from rasa_core.policies.memoization import MemoizationPolicy
-from rasa_core.policies.online_policy_trainer import (
-    OnlinePolicyTrainer)
 from rasa_core.processor import MessageProcessor
 from rasa_core.tracker_store import InMemoryTrackerStore, TrackerStore
+from rasa_core.trackers import DialogueStateTracker
 
 logger = logging.getLogger(__name__)
+
+if typing.TYPE_CHECKING:
+    from rasa_core.interpreter import NaturalLanguageInterpreter as NLI
 
 
 class Agent(object):
@@ -34,34 +38,47 @@ class Agent(object):
     def __init__(
             self,
             domain,  # type: Union[Text, Domain]
-            policies=None,  # type: Optional[Union[PolicyEnsemble, List[Policy]]
-            featurizer=None,  # type: Optional[Featurizer]
-            interpreter=None,  # type: Optional[NaturalLanguageInterpreter]
+            policies=None,  # type: Union[PolicyEnsemble, List[Policy], None]
+            interpreter=None,  # type: Union[NLI, Text, None]
             tracker_store=None  # type: Optional[TrackerStore]
     ):
         self.domain = self._create_domain(domain)
-        self.featurizer = self._create_featurizer(featurizer)
         self.policy_ensemble = self._create_ensemble(policies)
         self.interpreter = NaturalLanguageInterpreter.create(interpreter)
         self.tracker_store = self.create_tracker_store(
                 tracker_store, self.domain)
 
     @classmethod
-    def load(cls, path, interpreter=None, tracker_store=None,
-             action_factory=None):
+    def load(cls,
+             path,  # type: Text
+             interpreter=None,  # type: Union[NLI, Text, None]
+             tracker_store=None,  # type: Optional[TrackerStore]
+             action_factory=None  # type: Optional[Text]
+             ):
         # type: (Text, Any, Optional[TrackerStore]) -> Agent
+        """Load a persisted model from the passed path."""
 
         if path is None:
             raise ValueError("No domain path specified.")
+
+        if os.path.isfile(path):
+            raise ValueError("You are trying to load a MODEL from a file "
+                             "('{}'), which is not possible. \n"
+                             "The persisted path should be a directory "
+                             "containing the various model files. \n\n"
+                             "If you want to load training data instead of "
+                             "a model, use `agent.load_data(...)` "
+                             "instead.".format(path))
+
+        ensemble = PolicyEnsemble.load(path)
         domain = TemplateDomain.load(os.path.join(path, "domain.yml"),
                                      action_factory)
         # ensures the domain hasn't changed between test and train
         domain.compare_with_specification(path)
-        featurizer = Featurizer.load(path)
-        ensemble = PolicyEnsemble.load(path, featurizer)
         _interpreter = NaturalLanguageInterpreter.create(interpreter)
         _tracker_store = cls.create_tracker_store(tracker_store, domain)
-        return cls(domain, ensemble, featurizer, _interpreter, _tracker_store)
+
+        return cls(domain, ensemble, _interpreter, _tracker_store)
 
     def handle_message(
             self,
@@ -100,6 +117,7 @@ class Agent(object):
                                text_message,
                                sender_id=UserMessage.DEFAULT_SENDER_ID):
         # type: (Text, Optional[Text]) -> Dict[Text, Any]
+        """Start to process a messages, returning the next action to take. """
 
         processor = self._create_processor()
         return processor.start_message_handling(
@@ -107,6 +125,9 @@ class Agent(object):
 
     def continue_message_handling(self, sender_id, executed_action, events):
         # type: (Text, Text, List[Event]) -> Dict[Text, Any]
+        """Continue to process a messages.
+
+        Predicts the next action to take by the caller"""
 
         processor = self._create_processor()
         return processor.continue_message_handling(sender_id,
@@ -134,74 +155,153 @@ class Agent(object):
         training data."""
 
         for p in self.policy_ensemble.policies:
-            # explicitly ignore inheritance (e.g. scoring policy)
+            # explicitly ignore inheritance (e.g. augmented memoization policy)
             if type(p) == MemoizationPolicy:
                 p.toggle(activate)
 
-    def train(self, resource_name=None, model_path=None, remove_duplicates=True,
-              **kwargs):
-        # type: (Optional[Text], Optional[Text], **Any) -> None
-        """Train the policies / policy ensemble using dialogue data from file"""
+    def load_data(self,
+                  resource_name,  # type: Text
+                  remove_duplicates=True,  # type: bool
+                  augmentation_factor=20,  # type: int
+                  max_number_of_trackers=2000,  # type: int
+                  tracker_limit=None,  # type: Optional[int]
+                  use_story_concatenation=True  # type: bool
+                  ):
+        # type: (...) -> List[DialogueStateTracker]
+        """Load training data from a resource."""
 
-        trainer = PolicyTrainer(self.policy_ensemble, self.domain,
-                                self.featurizer)
-        trainer.train(resource_name, remove_duplicates=remove_duplicates,
-                      **kwargs)
+        return training.load_data(resource_name, self.domain, remove_duplicates,
+                                  augmentation_factor, max_number_of_trackers,
+                                  tracker_limit, use_story_concatenation)
 
-        if model_path:
-            self.persist(model_path)
+    def train(self,
+              training_trackers,  # type: List[DialogueStateTracker]
+              **kwargs  # type: **Any
+              ):
+        # type: (...) -> None
+        """Train the policies / policy ensemble using dialogue data from file.
+
+            :param training_trackers: trackers to train on
+            :param kwargs: additional arguments passed to the underlying ML
+                           trainer (e.g. keras parameters)
+        """
+
+        # deprecation tests
+        if kwargs.get('featurizer') or kwargs.get('max_history'):
+            raise Exception("Passing `featurizer` and `max_history` "
+                            "to `agent.train(...)` is not supported anymore. "
+                            "Pass appropriate featurizer "
+                            "directly to the policy instead. More info "
+                            "https://core.rasa.com/migrations.html#x-to-0-9-0")
+
+        # TODO: DEPRECATED - remove in version 0.10
+        if isinstance(training_trackers, string_types):
+            # the user most likely passed in a file name to load training
+            # data from
+            logger.warning("Passing a file name to `agent.train(...)` is "
+                           "deprecated. Rather load the data with "
+                           "`data = agent.load_data(file_name)` and pass it "
+                           "to `agent.train(data)`.")
+            training_trackers = self.load_data(training_trackers)
+
+        logger.debug("Agent trainer got kwargs: {}".format(kwargs))
+        check_domain_sanity(self.domain)
+
+        self.policy_ensemble.train(training_trackers, self.domain,
+                                   **kwargs)
 
     def train_online(self,
-                     resource_name=None,  # type: Optional[Text]
+                     training_trackers,  # type: List[DialogueStateTracker]
                      input_channel=None,  # type: Optional[InputChannel]
-                     model_path=None,  # type: Optional[Text]
+                     max_visual_history=3,  # type: int
                      **kwargs  # type: **Any
                      ):
         # type: (...) -> None
-        """Runs an online training session on the set policies / ensemble.
-
-        The policies will be pretrained using the data from `filename`.
-        After that the model will get trained on dialogues from the input
-        channel. During the dialogue the annotations and state of the agent
-        can be changed to correct wrong behaviour."""
+        from rasa_core.policies.online_trainer import OnlinePolicyEnsemble
+        """Train a policy ensemble in online learning mode."""
 
         if not self.interpreter:
             raise ValueError(
                     "When using online learning, you need to specify "
                     "an interpreter for the agent to use.")
-        trainer = OnlinePolicyTrainer(self.policy_ensemble, self.domain,
-                                      self.featurizer)
-        trainer.train(resource_name, self.interpreter, input_channel, **kwargs)
 
-        if model_path:
-            self.persist(model_path)
+        # TODO: DEPRECATED - remove in version 0.10
+        if isinstance(training_trackers, string_types):
+            # the user most likely passed in a file name to load training
+            # data from
+            logger.warning("Passing a file name to `agent.train_online(...)` "
+                           "is deprecated. Rather load the data with "
+                           "`data = agent.load_data(file_name)` and pass it "
+                           "to `agent.train_online(data)`.")
+            training_trackers = self.load_data(training_trackers)
+
+        logger.debug("Agent online trainer got kwargs: {}".format(kwargs))
+        check_domain_sanity(self.domain)
+
+        self.policy_ensemble.train(training_trackers, self.domain, **kwargs)
+
+        ensemble = OnlinePolicyEnsemble(self.policy_ensemble,
+                                        training_trackers,
+                                        max_visual_history)
+
+        ensemble.run_online_training(self.domain, self.interpreter,
+                                     input_channel)
+
+    @staticmethod
+    def _clear_model_directory(model_path):
+        # type: (Text) -> None
+        """Remove existing files from model directory.
+
+        Only removes files if the directory seems to contain a previously
+        persisted model. Otherwise does nothing to avoid deleting
+        `/` by accident."""
+
+        if not os.path.exists(model_path):
+            return
+
+        domain_spec_path = os.path.join(model_path, 'policy_metadata.json')
+        # check if there were a model before
+        if os.path.exists(domain_spec_path):
+            logger.info("Model directory {} exists and contains old "
+                        "model files. All files will be overwritten."
+                        "".format(model_path))
+            shutil.rmtree(model_path)
+        else:
+            logger.debug("Model directory {} exists, but does not contain "
+                         "all old model files. Some files might be "
+                         "overwritten.".format(model_path))
 
     def persist(self, model_path):
         # type: (Text) -> None
         """Persists this agent into a directory for later loading and usage."""
 
+        self._clear_model_directory(model_path)
+
         self.policy_ensemble.persist(model_path)
         self.domain.persist(os.path.join(model_path, "domain.yml"))
         self.domain.persist_specification(model_path)
-        self.featurizer.persist(model_path)
 
         logger.info("Persisted model to '{}'"
                     "".format(os.path.abspath(model_path)))
 
     def visualize(self,
-                  resource_name,
-                  output_file,
-                  max_history,
-                  nlu_training_data=None,
-                  fontsize=12
+                  resource_name,  # type: Text
+                  output_file,  # type: Text
+                  max_history,  # type: int
+                  nlu_training_data=None,  # type: Optional[Text]
+                  should_merge_nodes=True,  # type: bool
+                  fontsize=12  # type: int
                   ):
+        # type: (...) -> None
         from rasa_core.training.visualization import visualize_stories
-        from rasa_core.training import StoryFileReader
+        from rasa_core.training.dsl import StoryFileReader
+        """Visualize the loaded training data from the resource."""
 
         story_steps = StoryFileReader.read_from_folder(resource_name,
                                                        self.domain)
         visualize_stories(story_steps, self.domain, output_file, max_history,
-                          self.interpreter, nlu_training_data, fontsize)
+                          self.interpreter, nlu_training_data,
+                          should_merge_nodes, fontsize)
 
     def _ensure_agent_is_prepared(self):
         # type: () -> None
@@ -211,13 +311,12 @@ class Agent(object):
         Raises an exception if any argument is missing."""
 
         if self.interpreter is None or self.tracker_store is None:
-            raise Exception(
-                    "Agent needs to be prepared before usage. "
-                    "You need to set an interpreter as well "
-                    "as a tracker store.")
+            raise Exception("Agent needs to be prepared before usage. "
+                            "You need to set an interpreter as well "
+                            "as a tracker store.")
 
     def _create_processor(self, preprocessor=None):
-        # type: (Callable[[Text], Text]) -> MessageProcessor
+        # type: (Optional[Callable[[Text], Text]]) -> MessageProcessor
         """Instantiates a processor based on the set state of the agent."""
 
         self._ensure_agent_is_prepared()
@@ -225,12 +324,10 @@ class Agent(object):
                 self.interpreter, self.policy_ensemble, self.domain,
                 self.tracker_store, message_preprocessor=preprocessor)
 
-    @classmethod
-    def _create_featurizer(cls, featurizer):
-        return featurizer if featurizer is not None else BinaryFeaturizer()
+    @staticmethod
+    def _create_domain(domain):
+        # type: (Union[Domain, Text]) -> Domain
 
-    @classmethod
-    def _create_domain(cls, domain):
         if isinstance(domain, string_types):
             return TemplateDomain.load(domain)
         elif isinstance(domain, Domain):
@@ -241,8 +338,8 @@ class Agent(object):
                     "specification or a domain instance. But got "
                     "type '{}' with value '{}'".format(type(domain), domain))
 
-    @classmethod
-    def create_tracker_store(cls, store, domain):
+    @staticmethod
+    def create_tracker_store(store, domain):
         # type: (Optional[TrackerStore], Domain) -> TrackerStore
         if store is not None:
             store.domain = domain
@@ -252,15 +349,16 @@ class Agent(object):
 
     @staticmethod
     def _create_interpreter(
-            interp  # Optional[Union[Text, NaturalLanguageInterpreter]]
+            interp  # type: Union[Text, NaturalLanguageInterpreter, None]
     ):
         # type: (...) -> NaturalLanguageInterpreter
         return NaturalLanguageInterpreter.create(interp)
 
     @staticmethod
     def _create_ensemble(policies):
+        # type: (Union[List[Policy], PolicyEnsemble, None]) -> PolicyEnsemble
         if policies is None:
-            return SimplePolicyEnsemble([MemoizationPolicy])
+            return SimplePolicyEnsemble([])
         if isinstance(policies, list):
             return SimplePolicyEnsemble(policies)
         elif isinstance(policies, PolicyEnsemble):