+++ tests/test_training.py
@@ -6,7 +6,7 @@ from __future__ import unicode_literals
 from rasa_core.interpreter import RegexInterpreter
 from rasa_core.train import train_dialogue_model
 
-from rasa_core.training import StoryFileReader
+from rasa_core.training.dsl import StoryFileReader
 from rasa_core.training.visualization import visualize_stories
 from tests.conftest import DEFAULT_DOMAIN_PATH, DEFAULT_STORIES_FILE
 
@@ -16,17 +16,19 @@ def test_story_visualization_script():
     assert create_argument_parser() is not None
 
 
-def test_story_visualization(default_domain):
+def test_story_visualization(default_domain, tmpdir):
     story_steps = StoryFileReader.read_from_file(
             "data/test_stories/stories.md", default_domain,
             interpreter=RegexInterpreter())
+    out_file = tmpdir.join("graph.png").strpath
     generated_graph = visualize_stories(story_steps, default_domain,
-                                        output_file=None,
+                                        output_file=out_file,
                                         max_history=3,
                                         should_merge_nodes=False)
-    assert len(generated_graph.nodes()) == 33
 
-    assert len(generated_graph.edges()) == 36
+    assert len(generated_graph.nodes()) == 51
+
+    assert len(generated_graph.edges()) == 56
 
 
 def test_story_visualization_with_merging(default_domain):