+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ This project adheres to `Semantic Versioning`_ starting with version 0.2.0.
 
 .. note:: This version is not yet released and is under active development.
 
+This is a major new release with backward incompatible changes. Old trained
+models can not be read with the new version - you need to retrain your model.
+
 Added
 -----
 - supported loading training data from a folder - loads all stories from
@@ -33,6 +36,7 @@ Removed
 Fixed
 -----
 - Slack connector: ``slack_channel`` kwarg is used to send messages either back to the user or to a static channel
+- properly log to a file when using the ``run`` script
 
 [0.8.2] - 2018-02-13
 ^^^^^^^^^^^^^^^^^^^^